Refactor (core): Removed the EventScheduler module
The src/core/event_scheduler.py file has been removed. It contains the implementation of the intelligent event scheduler. This scheduler is responsible for intelligent event distribution and execution strategy selection, supporting multiple scheduling strategies and execution modes. This removal may be due to simplification of the system architecture or the use of alternative solutions.

Signed-off-by: Donny <donnymoving@gmail.com>